% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_S3.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Extract coefficients from ridgereg object}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ridgereg"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector of ridge regression coefficients.
}
\description{
Extract coefficients from ridgereg object
}
