% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_S3.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for ridgereg objects}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class \code{"ridgereg"}.}

\item{newdata}{Optional data frame for prediction. If omitted, fitted values are returned.}
}
\value{
A numeric vector of predicted values.
}
\description{
Predict method for ridgereg objects
}
