% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_S3.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print method for ridgereg objects}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ridgereg"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for ridgereg objects
}
