% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_S3.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression using Least Squares}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula object specifying the model. (class: formula)}

\item{data}{A data frame containing the variables in the formula.}

\item{lambda}{A non-negative numeric value specifying the ridge penalty parameter.}
}
\value{
An object of class \code{"ridgereg"} containing:
\item{formula}{The model formula.}
\item{coefficients}{A numeric vector of estimated ridge regression coefficients.}
\item{fitted_values}{A numeric vector of fitted values.}
}
\description{
Computes ridge regression coefficients and fitted values based on a given formula, dataset, and regularization parameter, lambda.
The predictors are normalized before fitting the model.
}
